refactor(RateLimitService): moved email rate limit service to their own file
